requirements.txt
streamlit
langchain
openai
psycopg2-binary
python-dotenv
pandas
plotly
tabulate
